requirements.txt
git+https://github.com/Rapptz/discord.py
PyNaCl==1.3.0
async-timeout
python-dotenv